--- PlantNurserySystem/src/com/amdocs/plant/exception/CountryNotFoundException.java
@@ -1,7 +1,0 @@
-package com.amdocs.plant.exception;
-
-public class CountryNotFoundException extends Exception{
-	public String toString() {
-		return ("CountryNotFoundException occured.");
-	}
-}
