+++ src/com/amdocs/plant/dao/DBConnection.java
@@ -1,0 +1,20 @@
+package com.amdocs.plant.dao;
+
+import java.sql.*;
+
+public class DBConnection {
+	private Connection con;
+	
+	public DBConnection() {
+		try {
+			con = DriverManager.getConnection("Jdbc:Oracle:thin:@localhost:1521:orcl", "scott", "tiger"); // connection
+
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+	}
+
+	public Connection getCon() {
+		return con;
+	}
+}
