+++ src/com/amdocs/plant/exception/InvalidCostException.java
@@ -1,0 +1,7 @@
+package com.amdocs.plant.exception;
+
+public class InvalidCostException  extends Exception{
+	public String toString() {
+		return ("InvalidCostException occured.");
+	}
+}
