+++ src/com/amdocs/plant/exception/PlantIDNotFoundException.java
@@ -1,0 +1,7 @@
+package com.amdocs.plant.exception;
+
+public class PlantIDNotFoundException extends Exception {
+	public String toString() {
+		return ("PlantIDNotFoundException occured.");
+	}
+}
